+++ offline_setup.sh
@@ -74,6 +74,7 @@ REPLACE="
 /system/system_ext/priv-app/ArrowLauncher
 /system/system_ext/priv-app/ThemePicker
 /system/system_ext/priv-app/WallpaperPickerGoogleRelease
+/system/system_ext/priv-app/ParanoidThemePicker
 /system/product/overlay/ThemedIconsOverlay.apk
 /system/product/overlay/PixelLauncherIconsOverlay.apk
 /system/product/overlay/CustomPixelLauncherOverlay.apk