+++ changelog.md
@@ -4,6 +4,7 @@
 - Removed push notifications service
 - Fixed crashing issue while tapping on Google Search Bar
 - Added new function to download files in online installer
+- Fixed crashing issue when user installs AOSP Wallpaper Picker in Android 13 QPR2
 
 **Version 1.5**
 - Added theme switcher background & updated icons in extended settings app