+++ changelog.md
@@ -1,3 +1,7 @@
+**Version 1.6**
+- Fixed banner not showing while making launcher using PLE Builder
+- Fixed wallpaper app not opening in paranoid rom
+
 **Version 1.5**
 - Added theme switcher background & updated icons in extended settings app
 - Added Icon Shape Changer app